requirements.txt
requests==2.32.4
python-dotenv==1.1.1
numpy
flask
openai==1.97.0
anthropic==0.57.1
httpx<0.28  # Prevent incompatible version that breaks openai client
jsonschema>=4.0.0  # For sport configuration validation

# Testing dependencies
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-flask==1.3.0